feat: 대시보드 관련 컴포넌트 위치 변경
