+++ frontend/src/components/User/TypeTest/DevAnalysis.css
